# Add AVX512 GFNI processing