… .github/workflows/go.yml
… README.md
… galois.go
… galoisAvx512_amd64.go
… galoisAvx512_amd64.s
… galoisAvx512_amd64_test.go
… galois_gen_amd64.go
… galois_gen_amd64.s
… galois_gen_switch_amd64.go